split into modules and lint
