+++ .github/workflows/ci.yml
@@ -35,7 +35,7 @@ jobs:
       - name: Install poetry
         run: pip install "poetry>=1.1.8,<1.2"
       - name: Install dependencies
-        run: poetry install --extras "lint test"
+        run: poetry install
       - name: Lint
         run: poetry run flake8
       - name: Test
@@ -53,7 +53,7 @@ jobs:
     - name: Install poetry
       run: pip install "poetry>=1.1.8,<1.2"
     - name: Install dependencies
-      run: poetry install --extras "release"
+      run: poetry install
     - name: Build wheel
       run: poetry build
     - name: Twine check
