+++ pyproject.toml
@@ -11,21 +11,15 @@ readme = "README.md"
 python = ">=3.6,<3.10"
 
 [tool.poetry.dev-dependencies]
-black = { version = "*", optional = true }
-flake8 = { version = "*", optional = true }
-flake8-black = { version = "*", optional = true }
-flake8-import-order = { version = "*", optional = true }
-flake8-print = { version = "*", optional = true }
-pytest = { version = "*", optional = true }
-pytest-cov = { version = "*", optional = true }
-PySide6 = { version = "*", optional = true }
-twine = { version = "*", optional = true }
-
-[tool.poetry.extras]
-lint = ["black", "flake8", "flake8-black", "flake8-import-order", "flake8-print"]
-test = ["pytest", "pytest-cov"]
-examples = ["PySide6"]
-release = ["twine"]
+black = "*"
+flake8 = "*"
+flake8-black = "*"
+flake8-import-order = "*"
+flake8-print = "*"
+pytest = "*"
+pytest-cov = "*"
+PySide6 = "*"
+twine = "*"
 
 [build-system]
 requires = ["poetry>=0.12"]
