+++ setup.cfg
@@ -7,3 +7,5 @@ extend-ignore =
     E203,
 application-import-names = observ
 import-order-style = google
+per-file-ignores =
+    observ/__init__.py:F401,F403
